+++ NEWS.md
@@ -17,6 +17,8 @@ v1.0.0
 * Seamless (almost) launch of games without showing the desktop when starting and returning from RetroArch and other emulators
 * Per-game launch command override, so that different cores or emulators can be used on a per-game basis (saved to gamelist.xml)
 * Core location can be defined relative to the emulator binary using the %EMUPATH% varible in es_systems.cfg (mostly useful for Windows)
+* Properly implemented the option to show or hide hidden files and folders
+* Properly implemented the option to show or hide games flagged as hidden in the metadata editor
 * Help system updated and expanded to the complete application (previously it was only partially implemented)
 * Improved input device configuration, and default keyboard mappings are now applied if the keyboard has not been configured by the user
 * GUI-configurable option to sort favorite games on the top of the game lists (favorites marked with stars)